test(benchmarks) shows diff between std and loadr
Brings the benchmark between std library and
loadr in line with each other. The core performance
difference is how the std template.Execute handled unknown
data types and peforms reflection which has large
impact on performance when large data is present.

The idea is to allow for rendering with and without
base data.